Update ruby version and repair migrations
A) Ruby was at 2.4.5. Changed version and ran bundle install with no issues.

B) A migration was missing for actually creating the choruses table.
Not sure how that happened, but copied a migration before
the sequence break and changed it to be the new add choruses
migration. (Add 20190518171612_create_choruses.rb)